Create README.md
Hi Nabeel,

This is to inform you that we would like to take our discussion forward. As the next step, we would want you to work on the following assignment, and send it across within next 2-3 days. Do let me know if there're any questions / issues with the assignment. 

Assignment:

Build the front end for the web design given here - https://goo.gl/photos/sHFSdyUBXnagx3EA7. Make sure that the text boxes have following validations in place: 

1. PAN/TIN/Name : Alpha-numeric 
2. Date : Date format

On click of upload, open the explorer to select a dummy text file with PAN number, Name and Date, read it and fill the fields.

Every time the user selects upload, create a new folder in AWS S3 or Dropbox with time stamp as the name of the folder and store each text file using the APIs provided by AWS or Dropbox.

Thanks!